integrated all sub-programs
