+++ main.h
@@ -1,8 +1,8 @@
 #ifndef _MAIN_H
 #define _MAIN_H
 
-/* Structs */
 
+/* Structs */
 /**
  * struct list_s - singly linked list
  * @str: string - (malloc'ed string)
@@ -17,11 +17,12 @@ typedef struct list_s
 } list_t;
 
 
-/* Global Variables */
-extern char **environ; /* Enviroment Variable */
+/* Enviroment Variable */
+extern char **environ;
 
 
-/* Headers */
+/* External Header Files */
+#include <limits.h>
 #include <stdarg.h>
 #include <stddef.h>
 #include <stdio.h>
@@ -34,23 +35,27 @@ extern char **environ; /* Enviroment Variable */
 
 /* Function prototypes */
 void *add_node(const char *str);
-void free_list(list_t *head);
+int _atoi(const char *s);
+void chgdir(const char *path);
 void clear_env(void);
-int _unsetenv(const char *name);
-int _setenv(const char *name, const char *value, int overwrite);
-char *_strcpy(char *dest, const char *src);
-char *_strcat(char *dest, const char *src);
-
+int exe_bin(char **args);
 int exe_cmd(char **argv);
+void exit_status(const char *input);
+void free_list(list_t *head);
 int freeWords(char ***words, int wordCount);
-char *get_path(char *cmd, int *rt);
+char *get_path(char *cmd);
+int _setenv(const char *name, const char *value, int overwrite);
 int seperate_word(char *line,
 char ***words, int line_size, char *delim);
+char *_strcat(char *dest, const char *src);
 int _strcmp(const char *s1, const char *s2);
+char *_strcpy(char *dest, const char *src);
 char *_strdup(char *str);
 int _strlen(const char *s);
 char *_strtok(char *str, const char *delim);
 int output(char *s);
 int prompt(void);
+int _unsetenv(const char *name);
+
 
 #endif
